requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
flasgger==0.9.1
Flask==1.0.2
Flask-MySQL==1.4.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.3
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
numpy==1.16.2
plac==0.9.6
preshed==2.0.1
pycparser==2.18
PyMySQL==0.9.3
PySastrawi==1.2.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
regex==2017.11.9
requests==2.20.1
scikit-learn==0.19.2
scipy==1.2.1
six==1.11.0
thinc==6.12.0
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11